moved addStore/addIndex commands to services/rest project
