+++ services/rest/src/main/java/mil/nga/giat/geowave/service/rest/operations/AddAccumuloStoreCommand.java
@@ -8,7 +8,7 @@
  * Version 2.0 which accompanies this distribution and is available at
  * http://www.apache.org/licenses/LICENSE-2.0.txt
  ******************************************************************************/
-package mil.nga.giat.geowave.datastore.accumulo.operations;
+package mil.nga.giat.geowave.service.rest.operations;
 
 import java.io.File;
 import java.util.ArrayList;
